requirements.txt
Flask
psycopg[binary]
python-dotenv
requests
pandas
openpyxl
